Fix directory structure for gb
Needlessly nested
