+++ src/go-git-duet/git-duet-install-hook/main.go
