+++ src/go-git-duet/git-duet-pre-commit/main.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"time"
 
-	duet "github.com/jszwedko/go-git-duet"
+	"go-git-duet"
 )
 
 func main() {
