+++ src/go-git-duet/git-solo/main.go
@@ -6,7 +6,7 @@ import (
 
 	"code.google.com/p/getopt"
 
-	duet "github.com/jszwedko/go-git-duet"
+	"go-git-duet"
 )
 
 func main() {
@@ -40,20 +40,9 @@ func main() {
 			fmt.Println(err)
 			os.Exit(1)
 		}
-		committer, err := gitConfig.GetCommitter()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
 
 		printAuthor(author)
-		printCommitter(committer)
 		os.Exit(0)
-	}
-
-	if getopt.NArgs() != 2 {
-		fmt.Println("must specify two sets of initials")
-		os.Exit(1)
 	}
 
 	pairs, err := duet.NewPairsFromFile(configuration.PairsFile, configuration.EmailLookup)
@@ -67,24 +56,19 @@ func main() {
 		fmt.Println(err)
 		os.Exit(86)
 	}
+
 	if err = gitConfig.SetAuthor(author); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
 
-	committer, err := pairs.ByInitials(getopt.Arg(1))
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(86)
-	}
-	if err = gitConfig.SetCommitter(committer); err != nil {
+	if err = gitConfig.ClearCommitter(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
 
 	if !*quiet {
 		printAuthor(author)
-		printCommitter(committer)
 	}
 }
 
@@ -95,13 +79,4 @@ func printAuthor(author *duet.Pair) {
 
 	fmt.Printf("GIT_AUTHOR_NAME='%s'\n", author.Name)
 	fmt.Printf("GIT_AUTHOR_EMAIL='%s'\n", author.Email)
-}
-
-func printCommitter(committer *duet.Pair) {
-	if committer == nil {
-		return
-	}
-
-	fmt.Printf("GIT_COMMITTER_NAME='%s'\n", committer.Name)
-	fmt.Printf("GIT_COMMITTER_EMAIL='%s'\n", committer.Email)
 }
